# Always store state in an `Arc`