--- a/axum-extra/src/extract/query.rs
+++ b/axum-extra/src/extract/query.rs
@@ -62,7 +62,7 @@ impl<T, S, B> FromRequest<S, B> for Query<T>
 where
     T: DeserializeOwned,
     B: Send,
-    S: Send,
+    S: Send + Sync,
 {
     type Rejection = QueryRejection;