--- a/axum-macros/src/lib.rs
+++ b/axum-macros/src/lib.rs
@@ -226,7 +226,7 @@ mod typed_path;
 /// impl<S, B> FromRequest<S, B> for OtherExtractor
 /// where
 ///     B: Send,
-///     S: Send,
+///     S: Send + Sync,
 /// {
 ///     // this rejection doesn't implement `Display` and `Error`
 ///     type Rejection = (StatusCode, String);