--- a/axum/src/extract/content_length_limit.rs
+++ b/axum/src/extract/content_length_limit.rs
@@ -41,7 +41,7 @@ where
     T: FromRequest<S, B>,
     T::Rejection: IntoResponse,
     B: Send,
-    S: Send,
+    S: Send + Sync,
 {
     type Rejection = ContentLengthLimitRejection<T::Rejection>;