--- a/axum/src/extract/path/mod.rs
+++ b/axum/src/extract/path/mod.rs
@@ -167,7 +167,7 @@ impl<T, S, B> FromRequest<S, B> for Path<T>
 where
     T: DeserializeOwned + Send,
     B: Send,
-    S: Send,
+    S: Send + Sync,
 {
     type Rejection = PathRejection;