--- a/axum/src/json.rs
+++ b/axum/src/json.rs
@@ -100,7 +100,7 @@ where
     B: HttpBody + Send,
     B::Data: Send,
     B::Error: Into<BoxError>,
-    S: Send,
+    S: Send + Sync,
 {
     type Rejection = JsonRejection;