… axum-macros/tests/debug_handler/fail/extract_self_mut.rs
… axum-macros/tests/debug_handler/fail/extract_self_ref.rs
… axum-macros/tests/debug_handler/pass/result_impl_into_response.rs
… axum-macros/tests/debug_handler/pass/self_receiver.rs
… axum-macros/tests/from_request/pass/derive_opt_out.rs
… axum-macros/tests/from_request/pass/override_rejection.rs
… axum/src/docs/extract.md
… axum/src/routing/method_routing.rs
… axum/src/routing/mod.rs
… axum/src/typed_header.rs
… examples/consume-body-in-extractor-or-middleware/src/main.rs
… examples/customize-extractor-error/src/main.rs
… examples/customize-path-rejection/src/main.rs
… examples/jwt/src/main.rs
… examples/validator/src/main.rs
… examples/versioning/src/main.rs