Fixes https://github.com/tokio-rs/axum/issues/1269

Could remove the `S: Clone` bound but had to add `S: Sync` since `Arc<S>` is only `Send` if `S: Sync`.

## Benchmark results

### Baseline

```
Running "extension" benchmark
  Beginning round 1...
  Benchmarking 10 connections @ http://0.0.0.0:62118 for 10 second(s)
    Latencies:
      Avg      Stdev    Min      Max
      0.06ms   0.02ms   0.02ms   1.03ms
    Requests:
      Total: 1555665 Req/Sec: 155558.87
    Transfer:
      Total: 111.27 MB Transfer Rate: 11.13 MB/Sec
```

```
Running "state" benchmark
  Beginning round 1...
  Benchmarking 10 connections @ http://0.0.0.0:62129 for 10 second(s)
    Latencies:
      Avg      Stdev    Min      Max
      0.06ms   0.02ms   0.02ms   1.77ms
    Requests:
      Total: 1716178 Req/Sec: 171604.66
    Transfer:
      Total: 122.75 MB Transfer Rate: 12.27 MB/Sec
```

### With `Arc<S>`

```
Running "state" benchmark
  Beginning round 1...
  Benchmarking 10 connections @ http://0.0.0.0:62551 for 10 second(s)
    Latencies:
      Avg      Stdev    Min      Max
      0.06ms   0.02ms   0.01ms   0.73ms
    Requests:
      Total: 1744483 Req/Sec: 174435.79
    Transfer:
      Total: 124.78 MB Transfer Rate: 12.48 MB/Sec
```

So a relatively small increase of 1.6% but the state is only cloned once (and only if its actually extracted) instead of four times previously before the `Arc`.